[Feat-#16] Add google authentication
Add google strategy for authentication with google account,
bonus: add test comunication beteween frontend and backend to
get and store jwt token on client side for later usage.

The feature allows to login with google account in user with previous jwt strategy
that creates a valid token for user loged in with google account similar to
the regular local authentication.

for the react app just add some links and redirect function to access the
google consent screen to login with.