[Feat-#20] Add Product (partially) & Files saving
Add Product table with primitive fields
and images that saves the name of image in DB.
with function to create a Product.

Add File saving feature for saving images in "images" directory,
create a uniqe name for saved files with uuid.

+ Add to user referece for banned users table onDelete:'CASCADE' so if user were deleted
the BannedUser row for that user also deletes from DB.